requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
django-cleanup==7.0.0
mysqlclient==2.1.1
Pillow==9.5.0
sqlparse==0.4.3
tzdata==2023.3
